+++ config/peru-rules.php
@@ -1,8 +1,0 @@
-<?php
-
-// Opciones del paquete. Se publican con:
-//   php artisan vendor:publish --tag=peru-rules-config
-
-return [
-
-];